prevent reinstalls when running multiple times

+++ Azure/ARM/EventStore/installdup.sh
@@ -1,8 +1,14 @@
 #!/bin/bash
 
-sudo add-apt-repository ppa:duplicity-team/ppa
-sudo apt-get update
-sudo apt-get install duplicity
-sudo apt-get install python-pip
-sudo pip install azure-storage==0.20
-sudo pip install logentries
+
+if [ -f "/usr/bin/duplicity" ]
+then  
+    echo "Duplicity already installed"
+else
+    sudo add-apt-repository ppa:duplicity-team/ppa --yes
+    sudo apt-get update
+    sudo apt-get install duplicity --yes
+    sudo apt-get install python-pip --yes
+    sudo pip install azure-storage==0.20 
+    sudo pip install logentries 
+fi
